refactor: decompose ipc.ts god-module into per-domain registrars (#669)
ipc.ts was a 2,717-line god-module: all 187 ipcMain.handle registrations
in one registerIpcHandlers(), with 81 imports and embedded business logic
mixed in. It was the dominant merge-conflict surface and obscured which
channels belong to which subsystem.

Split into src/main/ipc/:
- helpers.ts — cross-domain helpers (winFromEvent, rootPathFromEvent,
  reindexFile, removeFromIndexes, listIndexableFiles, persistIndexes,
  broadcastRewritten, broadcastHeadingRename, the write-pipeline hooks).
- 17 register-<domain>.ts files, each exporting one register<Domain>()
  function with that domain's handlers moved VERBATIM (notebase, links,
  queries, graph, tags, templates, shell, git, sites, bibliography, tools,
  refactor, sources, compute, publish, conversation, bookmarks). Each
  imports only the subsystem it uses.
- Domain-specific embedded logic moved with its domain: the conversation
  system-prompt + claim-note builders into register-conversation; the
  compute-result formatting + ComputeProposal audit writer into
  register-compute.

ipc.ts is now a 37-line dispatcher: registerIpcHandlers() calls the 17
register functions. No behavior change — the set of registered channels is
byte-for-byte identical (verified) and all bodies are verbatim.

assert-safe-path-coverage.test.ts (#397) statically scanned ipc.ts to
verify path-taking handlers route through safe sinks; pointed it at the new
src/main/ipc/*.ts sources. Same assertions, same channels.

pnpm lint (0 errors) + full suite (2733 passed) green.
